refactor(test): split test_signal.cpp by DSP concern (P5 test-split)
test_signal.cpp had grown to 2,030 lines / 108 TEST_CASEs covering the
whole core/signal DSP subsystem. Following the established Phase-5
test-split pattern (test_canvas.cpp -> test_canvas_cg_paths.cpp), extract
three DSP-concern clusters into sibling test TUs, moving each TEST_CASE
block byte-verbatim with unchanged names, tags, and assertions:

- test_signal_filters.cpp  (368 lines, 17 cases) — Biquad, SVF,
  LadderFilter, LinkwitzRiley
- test_signal_spectral.cpp (294 lines, 14 cases) — WindowFunction, FFT,
  Convolver
- test_signal_meter.cpp    (228 lines, 12 cases) — MultiChannelMeter,
  MultiChannelBallistics

Each is registered as its own pulp_add_test_suite() target, mirroring
the existing pulp-test-signal wiring.

test_signal.cpp: 2,030 -> 1,170 lines (65 cases remaining: DelayLine,
Gain, SimpleMixer, Compressor, Limiter, SmoothedValue, ADSR,
Oversampler, Oscillator, WaveShaper, NoiseGate, Panner, Chorus, Phaser,
Reverb).

Pre-split total: 2,155 assertions in 108 test cases.
Post-split total: 2,155 assertions in 108 test cases (540 + 964 + 604 +
47 assertions; 65 + 17 + 14 + 12 cases). CTest discovery unchanged.
