+++ test/test_signal.cpp
@@ -561,93 +561,6 @@ TEST_CASE("ADSR retrigger restarts attack from a partial level",
     REQUIRE(after < 1.0f);
 }
 
-// ── Biquad ───────────────────────────────────────────────────────────────────
-
-TEST_CASE("Biquad lowpass attenuates high frequencies", "[signal][biquad]") {
-    Biquad filter;
-    filter.set_coefficients(Biquad::Type::lowpass, 200.0f, 0.707f, 44100.0f);
-
-    // Generate 10kHz sine (should be attenuated)
-    float sum_sq = 0;
-    for (int i = 0; i < 4410; ++i) {
-        float input = std::sin(2.0f * 3.14159f * 10000.0f * i / 44100.0f);
-        float output = filter.process(input);
-        if (i > 100) sum_sq += output * output; // Skip transient
-    }
-    float rms = std::sqrt(sum_sq / 4310.0f);
-    REQUIRE(rms < 0.01f); // Should be heavily attenuated
-}
-
-TEST_CASE("Biquad lowpass passes low frequencies", "[signal][biquad]") {
-    Biquad filter;
-    filter.set_coefficients(Biquad::Type::lowpass, 5000.0f, 0.707f, 44100.0f);
-
-    // Generate 100Hz sine (should pass through)
-    float sum_sq = 0;
-    for (int i = 0; i < 4410; ++i) {
-        float input = std::sin(2.0f * 3.14159f * 100.0f * i / 44100.0f);
-        float output = filter.process(input);
-        if (i > 200) sum_sq += output * output;
-    }
-    float rms = std::sqrt(sum_sq / 4210.0f);
-    REQUIRE(rms > 0.6f); // Should pass through mostly
-}
-
-TEST_CASE("Biquad highpass attenuates low frequencies", "[signal][biquad]") {
-    Biquad filter;
-    filter.set_coefficients(Biquad::Type::highpass, 5000.0f, 0.707f, 44100.0f);
-
-    float sum_sq = 0;
-    for (int i = 0; i < 4410; ++i) {
-        float input = std::sin(2.0f * 3.14159f * 100.0f * i / 44100.0f);
-        float output = filter.process(input);
-        if (i > 200) sum_sq += output * output;
-    }
-    float rms = std::sqrt(sum_sq / 4210.0f);
-    REQUIRE(rms < 0.05f);
-}
-
-TEST_CASE("Biquad reset clears state", "[signal][biquad]") {
-    Biquad filter;
-    filter.set_coefficients(Biquad::Type::lowpass, 1000.0f, 0.707f, 44100.0f);
-
-    filter.process(1.0f);
-    filter.process(0.5f);
-    filter.reset();
-
-    // After reset, should behave as if freshly initialized
-    float out = filter.process(0.0f);
-    REQUIRE_THAT(out, WithinAbs(0.0, 0.001));
-}
-
-TEST_CASE("Biquad coefficient variants process finite impulse buffers", "[signal][biquad]") {
-    struct Case {
-        Biquad::Type type;
-        float gain_db;
-    };
-
-    const std::array<Case, 6> cases{{
-        {Biquad::Type::bandpass, 0.0f},
-        {Biquad::Type::notch, 0.0f},
-        {Biquad::Type::allpass, 0.0f},
-        {Biquad::Type::peaking, 6.0f},
-        {Biquad::Type::low_shelf, 6.0f},
-        {Biquad::Type::high_shelf, -6.0f},
-    }};
-
-    for (const auto& c : cases) {
-        Biquad filter;
-        filter.set_coefficients(c.type, 1200.0f, 0.707f, 48000.0f, c.gain_db);
-
-        float impulse[] = {1.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f};
-        filter.process(impulse, 8);
-
-        for (float sample : impulse) {
-            REQUIRE(std::isfinite(sample));
-        }
-    }
-}
-
 // ── Oversampler ──────────────────────────────────────────────────────────────
 
 TEST_CASE("Oversampler dispatches x2 and x4 callback phases",
@@ -859,85 +772,6 @@ TEST_CASE("Oscillator zero frequency leaves phase stationary",
     osc.reset();
     REQUIRE_THAT(osc.phase(), WithinAbs(0.0f, 1e-6f));
     REQUIRE_THAT(osc.frequency(), WithinAbs(0.0f, 1e-6f));
-}
-
-// ── SVF ──────────────────────────────────────────────────────────────────────
-
-TEST_CASE("SVF lowpass attenuates high frequencies", "[signal][svf]") {
-    Svf filter;
-    filter.set_sample_rate(44100.0f);
-    filter.set_frequency(200.0f);
-    filter.set_resonance(0.707f);
-    filter.set_mode(Svf::Mode::lowpass);
-
-    float sum_sq = 0;
-    for (int i = 0; i < 4410; ++i) {
-        float input = std::sin(2.0f * 3.14159f * 10000.0f * i / 44100.0f);
-        float output = filter.process(input);
-        if (i > 100) sum_sq += output * output;
-    }
-    float rms = std::sqrt(sum_sq / 4310.0f);
-    REQUIRE(rms < 0.01f);
-}
-
-TEST_CASE("SVF highpass passes high frequencies", "[signal][svf]") {
-    Svf filter;
-    filter.set_sample_rate(44100.0f);
-    filter.set_frequency(200.0f);
-    filter.set_resonance(0.707f);
-    filter.set_mode(Svf::Mode::highpass);
-
-    float sum_sq = 0;
-    for (int i = 0; i < 4410; ++i) {
-        float input = std::sin(2.0f * 3.14159f * 5000.0f * i / 44100.0f);
-        float output = filter.process(input);
-        if (i > 200) sum_sq += output * output;
-    }
-    float rms = std::sqrt(sum_sq / 4210.0f);
-    REQUIRE(rms > 0.5f);
-}
-
-TEST_CASE("SVF covers bandpass notch buffer and reset paths",
-          "[signal][svf][issue-645]") {
-    Svf filter;
-    filter.set_sample_rate(48000.0f);
-    filter.set_frequency(1000.0f);
-    filter.set_resonance(0.9f);
-
-    filter.set_mode(Svf::Mode::bandpass);
-    float band = filter.process(1.0f);
-    REQUIRE(std::isfinite(band));
-
-    filter.set_mode(Svf::Mode::notch);
-    float notched = filter.process(1.0f);
-    REQUIRE(std::isfinite(notched));
-
-    float buffer[] = {1.0f, 0.5f, 0.0f, -0.5f};
-    filter.process(buffer, 4);
-    for (float sample : buffer) {
-        REQUIRE(std::isfinite(sample));
-    }
-
-    filter.reset();
-    filter.set_mode(Svf::Mode::lowpass);
-    float after_reset = filter.process(0.0f);
-    REQUIRE_THAT(after_reset, WithinAbs(0.0f, 1e-6f));
-}
-
-TEST_CASE("SVF ignores nonpositive buffer lengths",
-          "[signal][svf][coverage][phase3-batch756]") {
-    Svf filter;
-    filter.set_sample_rate(48000.0f);
-    filter.set_frequency(1000.0f);
-    filter.set_mode(Svf::Mode::bandpass);
-
-    float samples[] = {0.25f, -0.5f, 0.75f};
-    filter.process(samples, 0);
-    filter.process(samples, -3);
-
-    REQUIRE_THAT(samples[0], WithinAbs(0.25f, 1e-6f));
-    REQUIRE_THAT(samples[1], WithinAbs(-0.5f, 1e-6f));
-    REQUIRE_THAT(samples[2], WithinAbs(0.75f, 1e-6f));
 }
 
 // ── WaveShaper ───────────────────────────────────────────────────────────────
@@ -1334,697 +1168,3 @@ TEST_CASE("Reverb handles zero decay, damping clamp, dry mix, and reset",
     REQUIRE_THAT(out.right, WithinAbs(0.25f, 1e-6f));
 }
 
-// ── LadderFilter ─────────────────────────────────────────────────────────────
-
-TEST_CASE("LadderFilter lowpass behavior", "[signal][ladder]") {
-    LadderFilter ladder;
-    ladder.set_sample_rate(44100.0f);
-    ladder.set_frequency(200.0f);
-    ladder.set_resonance(0.3f);
-
-    // High frequency should be attenuated
-    float sum_sq = 0;
-    for (int i = 0; i < 4410; ++i) {
-        float input = std::sin(2.0f * 3.14159f * 10000.0f * i / 44100.0f);
-        float output = ladder.process(input);
-        if (i > 200) sum_sq += output * output;
-    }
-    float rms = std::sqrt(sum_sq / 4210.0f);
-    REQUIRE(rms < 0.05f); // 24dB/oct should heavily attenuate
-}
-
-TEST_CASE("LadderFilter resets buffer state and clamps resonance inputs",
-          "[signal][ladder][issue-645]") {
-    LadderFilter ladder;
-    ladder.set_sample_rate(48000.0f);
-    ladder.set_frequency(1200.0f);
-    ladder.set_resonance(-2.0f);
-
-    float impulse[] = {1.0f, 0.0f, 0.0f, 0.0f};
-    ladder.process(impulse, 4);
-    for (float sample : impulse) {
-        REQUIRE(std::isfinite(sample));
-    }
-
-    ladder.reset();
-    REQUIRE_THAT(ladder.process(0.0f), WithinAbs(0.0f, 1e-6f));
-
-    ladder.set_resonance(3.0f);
-    ladder.set_frequency(8000.0f);
-    for (int i = 0; i < 32; ++i) {
-        const float sample = (i == 0) ? 0.75f : 0.0f;
-        REQUIRE(std::isfinite(ladder.process(sample)));
-    }
-
-    float unchanged[] = {0.25f, -0.25f};
-    ladder.process(unchanged, 0);
-    REQUIRE_THAT(unchanged[0], WithinAbs(0.25f, 1e-6f));
-    REQUIRE_THAT(unchanged[1], WithinAbs(-0.25f, 1e-6f));
-}
-
-TEST_CASE("LadderFilter reset restores fresh-filter impulse response",
-          "[signal][ladder][coverage][phase3-batch756]") {
-    auto impulse_response = [] {
-        LadderFilter filter;
-        filter.set_sample_rate(48000.0f);
-        filter.set_frequency(1600.0f);
-        filter.set_resonance(0.65f);
-
-        std::array<float, 6> response{};
-        response[0] = filter.process(1.0f);
-        for (std::size_t i = 1; i < response.size(); ++i)
-            response[i] = filter.process(0.0f);
-        return response;
-    };
-
-    const auto fresh = impulse_response();
-
-    LadderFilter reused;
-    reused.set_sample_rate(48000.0f);
-    reused.set_frequency(1600.0f);
-    reused.set_resonance(0.65f);
-    for (int i = 0; i < 16; ++i)
-        reused.process(0.25f);
-
-    reused.reset();
-    REQUIRE_THAT(reused.process(1.0f), WithinAbs(fresh[0], 1e-6f));
-    for (std::size_t i = 1; i < fresh.size(); ++i) {
-        INFO("impulse sample " << i);
-        REQUIRE_THAT(reused.process(0.0f), WithinAbs(fresh[i], 1e-6f));
-    }
-}
-
-TEST_CASE("LadderFilter ignores nonpositive buffer lengths",
-          "[signal][ladder][coverage][phase3-batch756]") {
-    LadderFilter ladder;
-    ladder.set_sample_rate(44100.0f);
-    ladder.set_frequency(1200.0f);
-
-    float samples[] = {0.25f, -0.5f, 0.75f};
-    ladder.process(samples, 0);
-    ladder.process(samples, -8);
-
-    REQUIRE_THAT(samples[0], WithinAbs(0.25f, 1e-6f));
-    REQUIRE_THAT(samples[1], WithinAbs(-0.5f, 1e-6f));
-    REQUIRE_THAT(samples[2], WithinAbs(0.75f, 1e-6f));
-}
-
-// ── LinkwitzRiley ────────────────────────────────────────────────────────────
-
-TEST_CASE("LinkwitzRiley splits into low and high bands", "[signal][lr]") {
-    LinkwitzRiley lr;
-    lr.set_frequency(1000.0f, 44100.0f);
-
-    // 200Hz signal should be mostly in the low band
-    float low_energy = 0, high_energy = 0;
-    for (int i = 0; i < 4410; ++i) {
-        float input = std::sin(2.0f * 3.14159f * 200.0f * i / 44100.0f);
-        auto split = lr.process(input);
-        if (i > 200) { // Skip transient
-            low_energy += split.low * split.low;
-            high_energy += split.high * split.high;
-        }
-    }
-    REQUIRE(low_energy > high_energy * 10.0f); // Low band should dominate
-
-    // 5kHz signal should be mostly in the high band
-    lr.reset();
-    low_energy = 0; high_energy = 0;
-    for (int i = 0; i < 4410; ++i) {
-        float input = std::sin(2.0f * 3.14159f * 5000.0f * i / 44100.0f);
-        auto split = lr.process(input);
-        if (i > 200) {
-            low_energy += split.low * split.low;
-            high_energy += split.high * split.high;
-        }
-    }
-    REQUIRE(high_energy > low_energy * 10.0f); // High band should dominate
-}
-
-TEST_CASE("LinkwitzRiley reset clears history while preserving coefficients",
-          "[signal][lr][issue-645]") {
-    auto impulse_response = [](LinkwitzRiley& lr) {
-        std::vector<LinkwitzRiley::BandSplit> response;
-        response.reserve(24);
-        for (int i = 0; i < 24; ++i) {
-            response.push_back(lr.process(i == 0 ? 1.0f : 0.0f));
-        }
-        return response;
-    };
-
-    LinkwitzRiley fresh;
-    fresh.set_frequency(1200.0f, 48000.0f);
-    auto expected = impulse_response(fresh);
-
-    LinkwitzRiley reused;
-    reused.set_frequency(1200.0f, 48000.0f);
-    for (int i = 0; i < 128; ++i) {
-        auto split = reused.process((i % 3 == 0) ? 0.75f : -0.25f);
-        REQUIRE(std::isfinite(split.low));
-        REQUIRE(std::isfinite(split.high));
-    }
-
-    reused.reset();
-    auto actual = impulse_response(reused);
-
-    REQUIRE(actual.size() == expected.size());
-    for (std::size_t i = 0; i < actual.size(); ++i) {
-        REQUIRE_THAT(actual[i].low, WithinAbs(expected[i].low, 1e-6f));
-        REQUIRE_THAT(actual[i].high, WithinAbs(expected[i].high, 1e-6f));
-    }
-}
-
-TEST_CASE("LinkwitzRiley cutoff boundary processing stays finite",
-          "[signal][lr][issue-645]") {
-    for (float cutoff : {0.0f, 20.0f, 20000.0f, 24000.0f}) {
-        LinkwitzRiley lr;
-        lr.set_frequency(cutoff, 48000.0f);
-
-        for (int i = 0; i < 64; ++i) {
-            float input = (i == 0) ? 1.0f : ((i % 2 == 0) ? 0.125f : -0.125f);
-            auto split = lr.process(input);
-            INFO("cutoff=" << cutoff << " sample=" << i);
-            REQUIRE(std::isfinite(split.low));
-            REQUIRE(std::isfinite(split.high));
-        }
-    }
-}
-
-TEST_CASE("LinkwitzRiley retuning after active history stays finite",
-          "[signal][lr][coverage][phase3-github]") {
-    LinkwitzRiley lr;
-    lr.set_frequency(300.0f, 48000.0f);
-    for (int i = 0; i < 32; ++i) {
-        (void)lr.process((i % 2 == 0) ? 0.5f : -0.5f);
-    }
-
-    lr.set_frequency(2400.0f, 48000.0f);
-    for (int i = 0; i < 16; ++i) {
-        auto split = lr.process(i == 0 ? 1.0f : 0.0f);
-        REQUIRE(std::isfinite(split.low));
-        REQUIRE(std::isfinite(split.high));
-    }
-}
-
-// ── WindowFunction ───────────────────────────────────────────────────────────
-
-TEST_CASE("WindowFunction Hann", "[signal][window]") {
-    auto w = WindowFunction::generate(256, WindowFunction::Type::hann);
-    REQUIRE(w.size() == 256);
-    REQUIRE_THAT(w[0], WithinAbs(0.0, 0.001));
-    REQUIRE_THAT(w[128], WithinAbs(1.0, 0.001)); // Peak at center
-    REQUIRE_THAT(w[255], WithinAbs(0.0, 0.001));
-}
-
-TEST_CASE("WindowFunction Blackman", "[signal][window]") {
-    auto w = WindowFunction::generate(512, WindowFunction::Type::blackman);
-    REQUIRE(w.size() == 512);
-    REQUIRE(w[256] > 0.9f); // Near peak
-    REQUIRE(w[0] < 0.01f);  // Near zero at edges
-}
-
-TEST_CASE("WindowFunction apply", "[signal][window]") {
-    auto w = WindowFunction::generate(4, WindowFunction::Type::rectangular);
-    float buf[] = {1.0f, 2.0f, 3.0f, 4.0f};
-    WindowFunction::apply(buf, w);
-    // Rectangular window should not change values
-    REQUIRE_THAT(buf[0], WithinAbs(1.0, 0.001));
-    REQUIRE_THAT(buf[3], WithinAbs(4.0, 0.001));
-}
-
-TEST_CASE("WindowFunction covers hamming flat-top and kaiser branches", "[signal][window][issue-645]") {
-    auto hamming = WindowFunction::generate(5, WindowFunction::Type::hamming);
-    REQUIRE_THAT(hamming.front(), WithinAbs(0.08f, 1e-5f));
-    REQUIRE_THAT(hamming.back(), WithinAbs(0.08f, 1e-5f));
-    REQUIRE_THAT(hamming[2], WithinAbs(1.0f, 1e-5f));
-
-    auto flat_top = WindowFunction::generate(9, WindowFunction::Type::flat_top);
-    REQUIRE(flat_top[4] > 0.99f);
-    REQUIRE(std::abs(flat_top.front()) < 0.001f);
-
-    auto kaiser_default = WindowFunction::generate(5, WindowFunction::Type::kaiser);
-    auto kaiser_custom = WindowFunction::generate(5, WindowFunction::Type::kaiser, 6.0f);
-    REQUIRE_THAT(kaiser_default.front(), WithinAbs(kaiser_default.back(), 1e-6f));
-    REQUIRE(kaiser_default[2] > kaiser_default.front());
-    REQUIRE(kaiser_custom.front() < kaiser_default.front());
-
-    float shaped[] = {1.0f, 1.0f, 1.0f, 1.0f, 1.0f};
-    WindowFunction::apply(shaped, hamming);
-    REQUIRE_THAT(shaped[0], WithinAbs(0.08f, 1e-5f));
-    REQUIRE_THAT(shaped[2], WithinAbs(1.0f, 1e-5f));
-}
-
-// ── FFT ──────────────────────────────────────────────────────────────────────
-
-TEST_CASE("FFT forward/inverse round-trip", "[signal][fft]") {
-    Fft fft(256);
-
-    // Create a simple signal
-    std::vector<std::complex<float>> data(256);
-    for (int i = 0; i < 256; ++i)
-        data[i] = {std::sin(2.0f * 3.14159f * 10.0f * i / 256.0f), 0.0f};
-
-    auto original = data;
-
-    fft.forward(data.data());
-    fft.inverse(data.data());
-
-    // Should recover original signal
-    for (int i = 0; i < 256; ++i) {
-        REQUIRE(std::abs(data[i].real() - original[i].real()) < 0.001f);
-    }
-}
-
-TEST_CASE("FFT detects single frequency", "[signal][fft]") {
-    constexpr int N = 1024;
-    Fft fft(N);
-
-    // Generate 100Hz sine at 44100 sample rate
-    std::vector<std::complex<float>> data(N);
-    float freq = 100.0f;
-    float sr = 44100.0f;
-    for (int i = 0; i < N; ++i)
-        data[i] = {std::sin(2.0f * 3.14159f * freq * i / sr), 0.0f};
-
-    fft.forward(data.data());
-
-    // Find the peak bin
-    int peak_bin = 0;
-    float peak_mag = 0;
-    for (int i = 1; i < N / 2; ++i) {
-        float mag = std::abs(data[i]);
-        if (mag > peak_mag) { peak_mag = mag; peak_bin = i; }
-    }
-
-    // Expected bin: freq * N / sr = 100 * 1024 / 44100 ≈ 2.32
-    float expected_bin = freq * N / sr;
-    REQUIRE(std::abs(peak_bin - expected_bin) < 2.0f);
-}
-
-TEST_CASE("FFT magnitude_db", "[signal][fft]") {
-    constexpr int N = 256;
-    Fft fft(N);
-
-    std::vector<std::complex<float>> freq(N);
-    std::vector<float> mag_db(N / 2);
-
-    // DC component = 1.0 → 0 dB
-    freq[0] = {1.0f, 0.0f};
-    fft.magnitude_db(freq.data(), mag_db.data(), N / 2);
-    REQUIRE_THAT(mag_db[0], WithinAbs(0.0, 0.01));
-
-    // 0.01 → -40 dB
-    freq[0] = {0.01f, 0.0f};
-    fft.magnitude_db(freq.data(), mag_db.data(), 1);
-    REQUIRE_THAT(mag_db[0], WithinAbs(-40.0, 0.1));
-}
-
-TEST_CASE("FFT forward_real", "[signal][fft]") {
-    constexpr int N = 256;
-    Fft fft(N);
-
-    std::vector<float> input(N, 0.0f);
-    input[0] = 1.0f; // Impulse
-
-    std::vector<std::complex<float>> output(N);
-    fft.forward_real(input.data(), output.data());
-
-    // FFT of impulse: all bins should have magnitude 1
-    for (int i = 0; i < N; ++i) {
-        REQUIRE(std::abs(std::abs(output[i]) - 1.0f) < 0.01f);
-    }
-}
-
-TEST_CASE("FFT move preserves transform state", "[signal][fft][issue-645]") {
-    Fft original(8);
-    Fft moved(std::move(original));
-
-    REQUIRE(original.size() == 0);
-    REQUIRE(moved.size() == 8);
-
-    std::vector<std::complex<float>> data(8, {1.0f, 0.0f});
-    moved.forward(data.data());
-
-    REQUIRE_THAT(data[0].real(), WithinAbs(8.0f, 1e-4f));
-    REQUIRE_THAT(data[0].imag(), WithinAbs(0.0f, 1e-4f));
-    for (int i = 1; i < 8; ++i) {
-        REQUIRE(std::abs(data[i]) < 1e-4f);
-    }
-
-    Fft assigned;
-    assigned = std::move(moved);
-    REQUIRE(moved.size() == 0);
-    REQUIRE(assigned.size() == 8);
-
-    assigned.inverse(data.data());
-    for (const auto& sample : data) {
-        REQUIRE_THAT(sample.real(), WithinAbs(1.0f, 1e-4f));
-        REQUIRE_THAT(sample.imag(), WithinAbs(0.0f, 1e-4f));
-    }
-}
-
-TEST_CASE("FFT magnitude helpers handle silence and complex bins", "[signal][fft][issue-645]") {
-    Fft fft(8);
-    std::vector<std::complex<float>> freq = {
-        {0.0f, 0.0f},
-        {3.0f, 4.0f},
-        {-0.5f, 0.0f},
-        {0.0f, -2.0f},
-        {0.0f, 0.0f},
-        {0.0f, 0.0f},
-        {0.0f, 0.0f},
-        {0.0f, 0.0f},
-    };
-    std::vector<float> linear(4, 0.0f);
-    std::vector<float> db(4, 0.0f);
-
-    fft.magnitude(freq.data(), linear.data(), 4);
-    fft.magnitude_db(freq.data(), db.data(), 4);
-
-    REQUIRE_THAT(linear[0], WithinAbs(0.0f, 1e-6f));
-    REQUIRE_THAT(linear[1], WithinAbs(5.0f, 1e-6f));
-    REQUIRE_THAT(linear[2], WithinAbs(0.5f, 1e-6f));
-    REQUIRE_THAT(linear[3], WithinAbs(2.0f, 1e-6f));
-    REQUIRE_THAT(db[0], WithinAbs(-200.0f, 0.01f));
-    REQUIRE_THAT(db[1], WithinAbs(13.9794f, 0.01f));
-    REQUIRE_THAT(db[2], WithinAbs(-6.0206f, 0.01f));
-    REQUIRE_THAT(db[3], WithinAbs(6.0206f, 0.01f));
-}
-
-TEST_CASE("FFT forward_real preserves exact-bin conjugate symmetry", "[signal][fft][issue-645]") {
-    constexpr int N = 64;
-    constexpr int bin = 5;
-    Fft fft(N);
-    std::vector<float> input(N);
-    for (int i = 0; i < N; ++i) {
-        input[i] = std::sin(2.0f * 3.14159265358979323846f * bin * i / N);
-    }
-
-    std::vector<std::complex<float>> output(N);
-    std::vector<float> magnitude(N, 0.0f);
-    fft.forward_real(input.data(), output.data());
-    fft.magnitude(output.data(), magnitude.data(), N);
-
-    REQUIRE_THAT(magnitude[0], WithinAbs(0.0f, 1e-4f));
-    REQUIRE_THAT(magnitude[bin], WithinAbs(static_cast<float>(N) / 2.0f, 1e-3f));
-    REQUIRE_THAT(magnitude[N - bin], WithinAbs(static_cast<float>(N) / 2.0f, 1e-3f));
-    REQUIRE_THAT(output[bin].real(), WithinAbs(output[N - bin].real(), 1e-4f));
-    REQUIRE_THAT(output[bin].imag(), WithinAbs(-output[N - bin].imag(), 1e-4f));
-}
-
-// ── Convolver ────────────────────────────────────────────────────────────────
-
-TEST_CASE("Convolver with identity IR", "[signal][convolver]") {
-    Convolver conv;
-
-    // Identity IR: [1, 0, 0, ...]
-    float ir[] = {1.0f};
-    conv.load_ir(ir, 1, 64);
-
-    // Process a known signal
-    std::vector<float> input(256);
-    std::vector<float> output(256);
-    for (int i = 0; i < 256; ++i)
-        input[i] = std::sin(2.0f * 3.14159f * 440.0f * i / 44100.0f);
-
-    conv.process(input.data(), output.data(), 256);
-
-    // Output should match input (after initial latency of one block)
-    float max_error = 0;
-    for (int i = 64; i < 200; ++i) {
-        float error = std::abs(output[i] - input[i - 64]);
-        if (error > max_error) max_error = error;
-    }
-    REQUIRE(max_error < 0.01f);
-}
-
-TEST_CASE("Convolver with simple delay IR", "[signal][convolver]") {
-    Convolver conv;
-
-    // Delay IR: [0, 0, 0, 0, 1] — 4-sample delay
-    float ir[] = {0.0f, 0.0f, 0.0f, 0.0f, 1.0f};
-    conv.load_ir(ir, 5, 32);
-
-    // Send an impulse and collect output
-    std::vector<float> output(128, 0.0f);
-    output[0] = conv.process(1.0f);
-    for (int i = 1; i < 128; ++i)
-        output[i] = conv.process(0.0f);
-
-    // Should see the impulse delayed by 4 samples + block latency
-    float peak = 0;
-    int peak_pos = 0;
-    for (int i = 0; i < 128; ++i) {
-        if (std::abs(output[i]) > peak) {
-            peak = std::abs(output[i]);
-            peak_pos = i;
-        }
-    }
-    REQUIRE(peak > 0.9f); // Should find the delayed impulse
-    REQUIRE(peak_pos >= 4); // At least 4 samples delayed
-}
-
-TEST_CASE("Convolver reset clears buffered overlap", "[signal][convolver][issue-645]") {
-    Convolver conv;
-    float ir[] = {0.0f, 1.0f, 0.5f};
-    conv.load_ir(ir, 3, 8);
-
-    std::vector<float> impulse(24, 0.0f);
-    std::vector<float> output(24, 0.0f);
-    impulse[0] = 1.0f;
-    conv.process(impulse.data(), output.data(), static_cast<int>(output.size()));
-
-    bool produced_tail = false;
-    for (float sample : output) {
-        produced_tail = produced_tail || std::abs(sample) > 1e-4f;
-    }
-    REQUIRE(produced_tail);
-
-    conv.reset();
-    std::fill(output.begin(), output.end(), 99.0f);
-    std::fill(impulse.begin(), impulse.end(), 0.0f);
-    conv.process(impulse.data(), output.data(), static_cast<int>(output.size()));
-
-    for (float sample : output) {
-        REQUIRE_THAT(sample, WithinAbs(0.0f, 1e-4f));
-    }
-}
-
-// ── MultiChannelMeter ────────────────────────────────────────────────────────
-
-TEST_CASE("MultiChannelMeter emits peak RMS clipping and stereo correlation", "[signal][meter]") {
-    MultiChannelMeter meter;
-    meter.prepare(44100.0f, 2);
-
-    std::vector<float> left(441, 0.5f);
-    std::vector<float> right(441, 0.5f);
-    left[10] = 1.2f;
-
-    const float* channels[] = {left.data(), right.data()};
-    meter.process(channels, 2, static_cast<int>(left.size()));
-
-    const auto& snap = meter.snapshot();
-    REQUIRE(snap.num_channels == 2);
-    REQUIRE_THAT(snap.channels[0].peak, WithinAbs(1.2f, 0.001f));
-    REQUIRE(snap.channels[0].clipped);
-    REQUIRE_THAT(snap.channels[1].peak, WithinAbs(0.5f, 0.001f));
-    REQUIRE_THAT(snap.channels[1].rms, WithinAbs(0.5f, 0.001f));
-    REQUIRE_THAT(snap.correlation, WithinAbs(1.0f, 0.005f));
-    REQUIRE(std::isfinite(snap.channels[0].lufs_momentary));
-}
-
-TEST_CASE("MultiChannelMeter tracks negative correlation and integrated loudness", "[signal][meter]") {
-    MultiChannelMeter meter;
-    meter.prepare(1000.0f, 2);
-
-    std::vector<float> left(10, 0.25f);
-    std::vector<float> right(10, -0.25f);
-    const float* channels[] = {left.data(), right.data()};
-
-    for (int i = 0; i < 40; ++i) {
-        meter.process(channels, 2, static_cast<int>(left.size()));
-    }
-
-    const auto& snap = meter.snapshot();
-    REQUIRE(snap.num_channels == 2);
-    REQUIRE_THAT(snap.correlation, WithinAbs(-1.0f, 0.001f));
-    REQUIRE(std::isfinite(snap.lufs_integrated));
-    REQUIRE_THAT(snap.lufs_integrated, WithinAbs(-12.73f, 0.1f));
-}
-
-TEST_CASE("MultiChannelMeter prepare clears stale clip flags", "[signal][meter]") {
-    MultiChannelMeter meter;
-    meter.prepare(1000.0f, 1);
-
-    std::vector<float> clipped(10, 1.2f);
-    const float* clipped_channels[] = {clipped.data()};
-    meter.process(clipped_channels, 1, static_cast<int>(clipped.size()));
-    REQUIRE(meter.snapshot().channels[0].clipped);
-
-    meter.prepare(1000.0f, 1);
-
-    std::vector<float> clean(10, 0.25f);
-    const float* clean_channels[] = {clean.data()};
-    meter.process(clean_channels, 1, static_cast<int>(clean.size()));
-    REQUIRE_FALSE(meter.snapshot().channels[0].clipped);
-}
-
-TEST_CASE("MultiChannelMeter clamps channels and reset clears snapshot", "[signal][meter][issue-645]") {
-    MultiChannelMeter meter;
-    meter.prepare(100.0f, kMaxMeterChannels + 4);
-    REQUIRE(meter.snapshot().num_channels == kMaxMeterChannels);
-
-    std::array<std::array<float, 1>, kMaxMeterChannels> samples{};
-    std::array<const float*, kMaxMeterChannels> channels{};
-    for (int ch = 0; ch < kMaxMeterChannels; ++ch) {
-        samples[ch][0] = static_cast<float>(ch + 1) / 20.0f;
-        channels[ch] = samples[ch].data();
-    }
-
-    meter.process(channels.data(), kMaxMeterChannels, 1);
-    REQUIRE(meter.snapshot().num_channels == kMaxMeterChannels);
-    REQUIRE_THAT(meter.snapshot().channels[15].peak, WithinAbs(0.8f, 1e-6f));
-
-    meter.reset();
-    REQUIRE(meter.snapshot().num_channels == 0);
-}
-
-TEST_CASE("MultiChannelMeter silent stereo emits finite zero correlation", "[signal][meter][issue-645]") {
-    MultiChannelMeter meter;
-    meter.prepare(10.0f, 2);
-
-    std::vector<float> left(4, 0.0f);
-    std::vector<float> right(4, 0.0f);
-    const float* channels[] = {left.data(), right.data()};
-    meter.process(channels, 2, static_cast<int>(left.size()));
-
-    const auto& snap = meter.snapshot();
-    REQUIRE(snap.num_channels == 2);
-    REQUIRE_THAT(snap.correlation, WithinAbs(0.0f, 1e-6f));
-    REQUIRE(std::isinf(snap.channels[0].lufs_momentary));
-    REQUIRE(std::isinf(snap.lufs_integrated));
-}
-
-TEST_CASE("MultiChannelMeter waits for block threshold before emitting", "[signal][meter][issue-645]") {
-    MultiChannelMeter meter;
-    meter.prepare(1000.0f, 1);
-
-    std::vector<float> samples(5, 0.75f);
-    const float* channels[] = {samples.data()};
-    meter.process(channels, 1, static_cast<int>(samples.size()));
-
-    REQUIRE(meter.snapshot().num_channels == 1);
-    REQUIRE_THAT(meter.snapshot().channels[0].peak, WithinAbs(0.0f, 1e-6f));
-}
-
-TEST_CASE("MultiChannelMeter resets correlation accumulation window", "[signal][meter][issue-645]") {
-    MultiChannelMeter meter;
-    meter.prepare(10.0f, 2);
-
-    float left[] = {0.5f};
-    float right[] = {0.5f};
-    const float* channels[] = {left, right};
-    meter.process(channels, 2, 1);
-
-    REQUIRE_THAT(meter.snapshot().correlation, WithinAbs(1.0f, 1e-6f));
-}
-
-TEST_CASE("MultiChannelMeter clamps negative prepared channel counts",
-          "[signal][meter][issue-645]") {
-    MultiChannelMeter meter;
-    meter.prepare(100.0f, -3);
-
-    const auto& snap = meter.snapshot();
-    REQUIRE(snap.num_channels == 0);
-    REQUIRE_THAT(snap.channels[0].peak, WithinAbs(0.0f, 1e-6f));
-}
-
-TEST_CASE("MultiChannelMeter clamps process channel count to prepared channels",
-          "[signal][meter][issue-645]") {
-    MultiChannelMeter meter;
-    meter.prepare(100.0f, 1);
-
-    float left[] = {0.25f};
-    float ignored_right[] = {1.0f};
-    std::vector<const float*> channels;
-    channels.push_back(left);
-    channels.push_back(ignored_right);
-    channels.push_back(nullptr);
-
-    meter.process(channels.data(), static_cast<int>(channels.size()), 1);
-
-    const auto& snap = meter.snapshot();
-    REQUIRE(snap.num_channels == 1);
-    REQUIRE_THAT(snap.channels[0].peak, WithinAbs(0.25f, 1e-6f));
-    REQUIRE_THAT(snap.channels[1].peak, WithinAbs(0.0f, 1e-6f));
-    REQUIRE_FALSE(snap.channels[1].clipped);
-}
-
-TEST_CASE("MultiChannelMeter clamps negative process channel counts",
-          "[signal][meter][issue-645]") {
-    MultiChannelMeter meter;
-    meter.prepare(100.0f, 1);
-
-    float sample[] = {1.0f};
-    const float* channels[] = {sample};
-    meter.process(channels, -1, 1);
-
-    const auto& snap = meter.snapshot();
-    REQUIRE(snap.num_channels == 0);
-    REQUIRE_THAT(snap.channels[0].peak, WithinAbs(0.0f, 1e-6f));
-    REQUIRE_FALSE(snap.channels[0].clipped);
-}
-
-TEST_CASE("MultiChannelBallistics holds peaks and clip indicators", "[signal][meter]") {
-    MultiChannelBallistics ballistics;
-
-    MultiChannelMeterData data;
-    data.num_channels = 1;
-    data.channels[0].peak = 1.0f;
-    data.channels[0].rms = 0.5f;
-    data.channels[0].clipped = true;
-
-    ballistics.update(data, 0.016f);
-    REQUIRE(ballistics.num_channels == 1);
-    REQUIRE(ballistics.channels[0].display_peak > 0.9f);
-    REQUIRE(ballistics.channels[0].display_rms > 0.45f);
-    REQUIRE_THAT(ballistics.channels[0].held_peak, WithinAbs(1.0f, 0.001f));
-    REQUIRE(ballistics.channels[0].clip_indicator);
-
-    data.channels[0].peak = 0.0f;
-    data.channels[0].rms = 0.0f;
-    data.channels[0].clipped = false;
-
-    ballistics.update(data, 0.5f);
-    REQUIRE(ballistics.channels[0].clip_indicator);
-    REQUIRE_THAT(ballistics.channels[0].held_peak, WithinAbs(1.0f, 0.001f));
-
-    ballistics.update(data, 3.0f);
-    REQUIRE_FALSE(ballistics.channels[0].clip_indicator);
-
-    data.channels[0].clipped = true;
-    ballistics.update(data, 0.016f);
-    REQUIRE(ballistics.channels[0].clip_indicator);
-    ballistics.clear_clips();
-    REQUIRE_FALSE(ballistics.channels[0].clip_indicator);
-}
-
-TEST_CASE("MultiChannelBallistics clamps invalid snapshot channel counts",
-          "[signal][meter][issue-645]") {
-    MultiChannelBallistics ballistics;
-
-    MultiChannelMeterData data;
-    data.num_channels = -2;
-    ballistics.update(data, 0.01f);
-    REQUIRE(ballistics.num_channels == 0);
-
-    data.num_channels = kMaxMeterChannels + 3;
-    for (int ch = 0; ch < kMaxMeterChannels; ++ch) {
-        data.channels[ch].peak = static_cast<float>(ch + 1) / 20.0f;
-    }
-
-    ballistics.update(data, 0.01f);
-    REQUIRE(ballistics.num_channels == kMaxMeterChannels);
-    REQUIRE(ballistics.channels[0].display_peak > 0.0f);
-    REQUIRE(ballistics.channels[kMaxMeterChannels - 1].display_peak > 0.0f);
-}
